Create 02_streamers_line_numbers
Line Numbers
Write a program that reads a text file and inserts line numbers in front of each of its lines. The result should be written to another text file. Use StreamReader in combination with StreamWriter.
 
text.txt	output.txt	   
-I was quick to judge him, but it wasn't his fault.
-Is this some kind of joke?! Is it?
-Quick, hide here…It is safer.	Line 1: -I was quick to judge him, but it wasn't his fault.
Line 2: -Is this some kind of joke?! Is it?
Line 3: -Quick, hide here…It is safer.