Create 09-ListOfPredicates.cs
List of Predicates
Find all numbers in the range 1...N that are divisible by the numbers of a given sequence. On the first line, you will be given an integer N – which is the end of the range. On the second line, you will be given a sequence of integers which are the dividers. Use predicates/functions.
Examples
 
Input	Output	   
10
1 1 1 2	2 4 6 8 10	   
100
2 5 10 20	20 40 60 80 100