Create 10-PredicateParty.cs
  Predicate Party!
Ivancho’s parents are on a vacation for the holidays and he is planning an epic party at home. Unfortunately, his organizational skills are next to non-existent, so you are given the task to help him with the reservations.
On the first line, you receive a list with all the people that are coming. On the next lines, until you get the "Party!" command, you may be asked to double or remove all the people that apply to a given criteria. There are three different criteria: 
Everyone that has his name starting with a given string
Everyone that has a name ending with a given string
Everyone that has a name with a given length.
Finally, print all the guests who are going to the party separated by ", " and then add the ending "are going to the party!". If there are no guests going to the party print "Nobody is going to the party!". See the examples below:
Examples
 
 Input	Output	   
Pesho Misho Stefan
Remove StartsWith P
Double Length 5
Party!	Misho, Misho, Stefan are going to the party!	   
Pesho
Double StartsWith Pesh
Double EndsWith esho
Party!	Pesho, Pesho, Pesho, Pesho are going to the party!	   
Pesho
Remove StartsWith P
Party!	Nobody is going to the party!