Create 12_tri_function.cs
 TriFunction
Write a program that traverses a collection of names and returns the first name, whose sum of characters is equal to or larger than a given number N, which will be given on the first line. Use a function that accepts another function as one of its parameters. Start off by building a regular function to hold the basic logic of the program. Something along the lines of Func<string, int, bool>. Afterwards create your main function which should accept the first function as one of its parameters.
Examples
 
Input	Output	   
800
Qvor Qnaki Petromir Sadam	Petromir