Add files via upload
Line Numbers
Write a program that reads a text file and inserts line numbers in front of each of its lines and count all the letters and punctuation marks. The result should be written to another text file. Use the static class File.
Examples
 
text.txt	output.txt	   
-I was quick to judge him, but it wasn't his fault.
-Is this some kind of joke?! Is it?
-Quick, hide here. It is safer.	Line 1: -I was quick to judge him, but it wasn't his fault. (37)(4)
Line 2: -Is this some kind of joke?! Is it? (24)(4)
Line 3: -Quick, hide here. It is safer. (22)(4)