Add files via upload
Generic Count Method Strings
Create a method that receives as an argument a list of any type, that can be compared and an element of the given type. The method should return the count of elements that are greater than the value of the given element. Modify your Box class to support comparison by value of the stored data.

Input
 On the first line, you will receive n - the number of elements to add to the list.
 On the next n lines, you will receive the actual elements.
 On the last line, you will get the value of the element for comparison.
Output
Print the count of elements that are larger than the value of the given element.
Examples
 
Input	Output	   
3
aa
aaa
bb
aa	2	 
Generic Count Method Doubles
Use the description of the previous problem, but now, test your list of generic boxes with doubles.
Examples
 
Input	Output	   
3
7.13
123.22
42.78
7.55	2